pyproject.toml
[tool.poetry]
name = "PyClits"
version = "1.0.0"
description = "Analyzer of nonlinear data"
authors = ["Nikola Jajcay, Hynek Lavicka"]

[tool.poetry.dependencies]
python = ">= 3.7, < 4.0"
scipy = ">=1.3"
numpy = "^1.3"
mpmath = "^1.1.0"
pandas = ">= 1.0.0, < 1.1.0"
matplotlib = "^3.1.2"
pip = "^20.0.2"
pydevd = "^1.9.0"
pydevd-pycharm = "^201.3803.74"
pytest-pycharm = "^0.6.0"
sphinx = "^2.4.0"
sphinx-rtd_theme = "^0.4.3"
scikit-learn = "^0.24.1"
netCDF4 = "^1.5.6"
pathos = "^0.2.7"
geos = "^0.2.3"
cartopy = "^0.18.0"
cryptoxlib-aio = "^5.2.0"
requests = "^2.26.0"
mysql-connector-python = "^8.0.26"
SQLAlchemy = "^1.4.25"
PyMySQL = "^1.0.2"
ccxt = "^1.58.55"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
